refactor: move migration function out of index
