+++ src/index.tsx
@@ -13,11 +13,13 @@ import { RestfulProvider } from 'restful-react';
 
 import { Section } from 'lib/fetchers';
 import { SavedSection } from 'lib/hooks/useSavedCourses';
+import { migrateLocalStorage } from 'lib/utils/localStorageMigration';
 
 import { Mobile } from 'common/mobile';
 
 import reportWebVitals from './reportWebVitals';
 import { Routes } from './routes';
+
 import './index.css';
 
 export type OldCourse = {
@@ -71,38 +73,9 @@ if (process.env.NODE_ENV === 'production' || process.env.REACT_APP_ANALYTICS) {
   firebase.analytics();
 }
 
-// insert cat cry emoji here
-// this migrates the localStorage format to a new one.
-try {
-  const item = window.localStorage.getItem('user:saved_courses');
-  if (item) {
-    const parsedItem = JSON.parse(item) as OldCourse[];
-    if (parsedItem.some((c) => c.sections.length > 0)) {
-      logEvent('local_storage_migration', { coursesLength: parsedItem.length });
-      localStorage.setItem(
-        'user:saved_courses',
-        JSON.stringify(
-          parsedItem.map(({ subject, pid, code, term, selected, lecture, lab, tutorial, color }) => ({
-            subject,
-            pid,
-            code,
-            term,
-            selected,
-            lecture: lecture?.sectionCode,
-            lab: lab?.sectionCode,
-            tutorial: tutorial?.sectionCode,
-            color,
-          }))
-        )
-      );
-    }
-  }
-} catch (error) {
-  console.warn(`Error reading localStorage key “user:saved_courses”:`, error);
-  console.warn('Likely the format is fine.');
-}
-
 const searchClient = algoliasearch('CR92D3S394', '5477854d63b676fe021f8f83f5839a3a');
+
+migrateLocalStorage();
 
 ReactDOM.render(
   <React.StrictMode>
