+++ src/lib/utils/index.ts
@@ -2,3 +2,4 @@ export * from './dates';
 export * from './terms';
 export * from './courses';
 export * from './constants';
+export * from './localStorageMigration';
